iboot: HS: Add support to build a signed iboot and embed it
This patch helps give the user an option to build a signed image
from within omapboot itself. The idea is to provide a single stage
build with an option to sign the binaries for HS/HD devices.

When the user passes in a path to the where the mshield is located,
the Makefile will use it to sign iboot and eboot for the user.

Example:
make CROSS_COMPILE=<path>/arm-eabi-4.4.3/bin/arm-eabi- ARCH=arm
MACH=omap5 BOARD=omap5uevm MSHIELD=<path>/omap-mshield-lite

Change-Id: Ifd58a93e3eed10bd3b05364238eb2c1736e2d191
Signed-off-by: Christina Warren <cawarren@ti.com>